Allow multiple classes in At.class' attribute values (closes #53).
`At.class'` ultimately calls `HTMLElement.classList.add` which means
that each call to `At.class'` can only contain one class (DOMToken).

Outside of the inconvenience (it is often the case that you want to
define multiple classes at once e.g. in CSS framework), it also causes a
weird discrepancy: `At.class'` is the only attribute for which Brr
reject values that would be valid for the corresponding HTML attribute.

This patch uses a similar implementation for `At.class'` as the
implementation for `At.style`, where the class fragments are accumulated
then joined to be passed to `setAttribute` at the end of attribute
processing.

This does not impact `El.class'` or `El.set_class`.